+++ src/main/scala/epic/slab/Slab.scala
@@ -40,4 +40,14 @@ object Slab {
     def substring(span: Span): String = substring(span.begin, span.end)
     def substring(begin: Int, end: Int): String = slab.content.substring(begin, end)
   }
+
+  implicit final class SpanIndexSlabOps[In <: HList](slab: Slab[String, In]) {
+    def covered[T <: SpanAnnotation](span: Span)(implicit sel: Selector[In, Vector[T]]): Iterable[T] = {
+      spanIndex(sel).apply(span)
+    }
+    def spanIndex[T <: SpanAnnotation](implicit sel: Selector[In, Vector[T]]) = SpanIndex(slab.select[T](sel))
+    def covered[T <: SpanAnnotation](span: Span, index: SpanIndex[T]): Iterable[T] = {
+      index(span)
+    }
+  }
 }
